refactor horizons api parsing logic
