+++ scripts/horizons-api/parse-horizons-data.js
@@ -1,22 +1,31 @@
-import * as fs from "node:fs";
-import * as path from "node:path";
-import { fileURLToPath } from "node:url";
-import { readSvsMoonInfoSync } from "./read-svs-moon-info.js";
+import { readSvsMoonInfoSync } from "../nasa-svs/read-svs-moon-info.js";
 import { fetchHorizonsData } from "./fetch.js";
 import { getCsvLines, parseLine, collateParsedLines } from "./parsers.js";
+import { relativePath, writeJsonFileSync } from "../file/index.js";
 
-const scriptDirPath = path.dirname(fileURLToPath(import.meta.url));
-const dataDirPath = path.resolve(scriptDirPath, "..", "data");
+async function parseData(year) {
+  const data = await fetchHorizonsData(`${year - 1}-12-31`, `${year}-12-31`);
+  const lines = data.toString().split("\n");
 
-const data = await fetchHorizonsData(`${year - 1}-12-31`, `${year}-12-31`);
-const lines = data.toString().split("\n");
+  const svsInfo = readSvsMoonInfoSync(year);
+  const parsedLines = getCsvLines(lines).map((line) =>
+    parseLine(line, svsInfo)
+  );
+  return collateParsedLines(parsedLines);
+}
 
-const svsInfo = readSvsMoonInfoSync(year);
-const parsedLines = getCsvLines(lines).map((line) => parseLine(line, svsInfo));
-const collatedLines = collateParsedLines(parsedLines);
+(async function main() {
+  const [_bin, _script, year] = process.argv;
+  if (!year) {
+    console.log("Please provide year");
+  }
+  const parsed = parseData(year);
+  const outputFilePath = relativePath(
+    import.meta.url,
+    "..",
 
-const outputFilePath = path.resolve(
-  dataDirPath,
-  `${year}-shonanko-horizons-svs.json`
-);
-fs.writeFileSync(outputFilePath, JSON.stringify(collatedLines, null, 2));
+    "data"`${year}-shonanko-horizons-svs.json`
+  );
+
+  writeJsonFileSync(outputFilePath, parsed);
+})();
